@@ -6,75 +6,72 @@
   "license": "MIT",
   "scripts": {
     "build": "cross-env NODE_ENV=production webpack",
-    "start": "cross-env NODE_ENV=development webpack serve"
+    "start": "cross-env NODE_ENV=development webpack serve",
+    "prepare": "husky"
   },
   "dependencies": {
-    "@babel/runtime": "^7.21.0",
-    "@fortawesome/fontawesome-svg-core": "^6.3.0",
-    "@fortawesome/free-solid-svg-icons": "^6.3.0",
+    "@babel/runtime": "^7.23.9",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "bootstrap": "^5.2.3",
-    "bootswatch": "^5.2.3",
-    "classnames": "^2.3.2",
+    "bootstrap": "^5.3.2",
+    "bootswatch": "^5.3.2",
+    "classnames": "^2.5.1",
     "convert-units": "3.0.0-beta.5",
-    "core-js": "^3.28.0",
-    "formik": "^2.2.9",
+    "core-js": "^3.35.1",
+    "formik": "^2.4.5",
     "lodash-es": "^4.17.21",
-    "normalize-scss": "^7.0.1",
+    "normalize-scss": "^8.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.7.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.1",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^6.8.1",
-    "reselect": "^4.1.7",
-    "use-local-storage-state": "^17.2.0"
+    "react-router-dom": "^6.22.0",
+    "reselect": "^5.1.0",
+    "use-local-storage-state": "^19.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.0",
-    "@babel/eslint-parser": "^7.19.1",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
-    "@babel/plugin-transform-regenerator": "^7.20.5",
-    "@babel/plugin-transform-runtime": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/register": "^7.21.0",
-    "autoprefixer": "^10.4.13",
-    "babel-loader": "^9.1.2",
+    "@babel/core": "^7.23.9",
+    "@babel/eslint-parser": "^7.23.10",
+    "@babel/preset-env": "^7.23.9",
+    "@babel/preset-react": "^7.23.3",
+    "@babel/register": "^7.23.7",
+    "autoprefixer": "^10.4.17",
+    "babel-loader": "^9.1.3",
     "clean-webpack-plugin": "^4.0.0",
     "cname-webpack-plugin": "^3.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.3",
-    "css-minimizer-webpack-plugin": "^4.2.2",
-    "eslint": "^8.34.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-import-resolver-webpack": "^0.13.2",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.32.2",
+    "css-loader": "^6.10.0",
+    "css-minimizer-webpack-plugin": "^6.0.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-webpack": "^0.13.8",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-webpack-plugin": "^4.0.0",
-    "html-loader": "^4.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^8.0.3",
-    "lint-staged": "^13.1.2",
-    "mini-css-extract-plugin": "^2.7.2",
-    "postcss": "^8.4.21",
+    "eslint-webpack-plugin": "^4.0.1",
+    "html-loader": "^5.0.0",
+    "html-webpack-plugin": "^5.6.0",
+    "husky": "^9.0.10",
+    "lint-staged": "^15.2.2",
+    "mini-css-extract-plugin": "^2.8.0",
+    "postcss": "^8.4.35",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^7.0.2",
-    "postcss-scss": "^4.0.6",
-    "prettier": "^2.8.4",
-    "sass": "^1.58.3",
-    "sass-loader": "^13.2.0",
-    "style-loader": "^3.3.1",
-    "stylelint": "^15.2.0",
-    "stylelint-config-recommended": "^10.0.1",
-    "stylelint-webpack-plugin": "^4.1.0",
-    "terser-webpack-plugin": "^5.3.6",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1"
+    "postcss-loader": "^8.1.0",
+    "postcss-scss": "^4.0.9",
+    "prettier": "^3.2.5",
+    "sass": "^1.70.0",
+    "sass-loader": "^14.1.0",
+    "style-loader": "^3.3.4",
+    "stylelint": "^16.2.1",
+    "stylelint-config-recommended": "^14.0.0",
+    "stylelint-webpack-plugin": "^5.0.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "webpack": "^5.90.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
